fix: Modified logic about open the modal
